Reorganized the functions and main function script written.
Shit won't work, cuz the hero file uses a different system (of print instead of using curses, and of creating its own enemy instead of using the enemy file).
Would be an easy fix, though this project is dead, so that won't happen.
I reorganized into an upkeep function to call when clearing the window, and a combat prep function for setting of variables relating to the importet enemy file, and I moved some stuff (pertaining to the creation of the screen) to main, and started on the player input for combat.